Change training run allocation id type from Long to Integer

+++ src/main/java/cz/muni/ics/kypo/training/adaptive/domain/training/TrainingRun.java
@@ -131,7 +131,7 @@ public class TrainingRun implements Serializable {
     @Column(name = "sandbox_instance_ref_id", length = 36)
     private String sandboxInstanceRefId;
     @Column(name = "sandbox_instance_allocation_id")
-    private Long sandboxInstanceAllocationId;
+    private Integer sandboxInstanceAllocationId;
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", nullable = false)
     private User participantRef;
@@ -287,7 +287,7 @@ public class TrainingRun implements Serializable {
      *
      * @return the sandbox instance allocation id
      */
-    public Long getSandboxInstanceAllocationId() {
+    public Integer getSandboxInstanceAllocationId() {
         return sandboxInstanceAllocationId;
     }
 
@@ -295,7 +295,7 @@ public class TrainingRun implements Serializable {
      * Sets sandbox instance allocation id associated with Training run
      * @param sandboxInstanceAllocationId the sandbox instance allocation id
      */
-    public void setSandboxInstanceAllocationId(Long sandboxInstanceAllocationId) {
+    public void setSandboxInstanceAllocationId(Integer sandboxInstanceAllocationId) {
         this.sandboxInstanceAllocationId = sandboxInstanceAllocationId;
     }
